Address M68K review comments
